#1 adiconar alunos as salas

+++ database/migrations/2019_07_12_173217_create_sala_user_table.php
@@ -15,10 +15,11 @@ class CreateSalaUserTable extends Migration
     {
         Schema::create('sala_user', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('sala_id')->unsigned();
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('sala_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->timestamps();
         });
+
     }
 
     /**
